+++ runtime-modules/referendum/src/benchmarking.rs
@@ -382,7 +382,7 @@ benchmarks_instance! {
     }
 
     on_initialize_revealing {
-        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1) as u32;
+        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1);
 
         let cycle_id = 1;
         let salt = vec![0u8];
@@ -499,7 +499,7 @@ benchmarks_instance! {
     }
 
     reveal_vote_space_for_new_winner {
-        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1) as u32;
+        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1);
 
         let salt = vec![0u8];
         let vote_option = 2 * (i + 1); // Greater than number of voters + number of candidates
@@ -563,7 +563,7 @@ benchmarks_instance! {
     }
 
     reveal_vote_space_not_in_winners {
-        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1) as u32;
+        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1);
 
         let salt = vec![0u8];
         let vote_option = 2 * (i + 1); // Greater than number of voters + number of candidates
@@ -617,7 +617,7 @@ benchmarks_instance! {
     }
 
     reveal_vote_space_replace_last_winner {
-        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1) as u32;
+        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1);
 
         let salt = vec![0u8];
         let vote_option = 2 * (i + 1); // Greater than number of voters + number of candidates
@@ -679,7 +679,7 @@ benchmarks_instance! {
     }
 
     reveal_vote_already_existing {
-        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1) as u32;
+        let i in 0 .. (T::MaxWinnerTargetCount::get() - 1);
 
         let salt = vec![0u8];
         let vote_option = i;
